fixing build (#1067)
fix build with parallel tests
