+++ .github/workflows/build.yml
@@ -11,6 +11,8 @@ jobs:
     name: Build
     runs-on: buildjet-8vcpu-ubuntu-2204
     steps:
+      - run: echo "$(du -sh $(ls -A ~) | sort -hr)"
+
       - name: Check if organization member
         id: is_organization_member
         uses: JamesSingleton/is-organization-member@1.0.1
@@ -49,15 +51,37 @@ jobs:
         uses: actions/checkout@v3.3.0
         with:
           fetch-depth: 0
-      #        with:
-      #          ref: ${{ github.event.pull_request.head.sha }}
 
+      - name: Setup Gradle # To cache ~/.gradle
+        uses: gradle/gradle-build-action@v2
 
-      - name: Cache Gradle
+      - name: Cache SonarQube packages
+        uses: actions/cache@v3
+        if: ${{ github.actor != 'dependabot[bot]' }}
+        with:
+          path: ~/.sonar/cache
+          key: ${{ runner.os }}-sonar
+
+      - name: Cache Frontend dependencies
         uses: actions/cache@v3
         with:
-          path: ~/.gradle
-          key: ${{ runner.os }}-gradle
+          path: frontend/node_modules
+          key: ${{ runner.os }}-frontend-${{ hashFiles('frontend/package-lock.json')}}
+          restore-keys: ${{ runner.os }}-frontend
+
+      - name: Cache Giskard test resources
+        uses: actions/cache@v3
+        with:
+          path: ~/.giskard
+          key: ${{ runner.os }}-giskard-test-resources-${{ hashFiles('python-client/tests/fixtures/**/*py')}}
+          restore-keys: ${{ runner.os }}-giskard-test-resources
+
+      - name: Cache Python dependencies
+        uses: actions/cache@v3
+        with:
+          path: python-client/.venv
+          key: ${{ runner.os }}-client-${{ hashFiles('python-client/pdm.lock')}}
+          restore-keys: ${{ runner.os }}-client
 
       - uses: actions/setup-node@v3
         with:
@@ -79,34 +103,6 @@ jobs:
         run: |
           python -m pip install virtualenv -U --upgrade pip
           echo "Python ($(which python)) packages:\n $(python -m pip list)"
-
-      - name: Cache SonarQube packages
-        uses: actions/cache@v3
-        if: ${{ github.actor != 'dependabot[bot]' }}
-        with:
-          path: ~/.sonar/cache
-          key: ${{ runner.os }}-sonar
-
-      - name: Cache Frontend dependencies
-        uses: actions/cache@v3
-        with:
-          path: frontend/node_modules
-          key: ${{ runner.os }}-frontend-${{ hashFiles('frontend/package-lock.json')}}
-          restore-keys: ${{ runner.os }}-frontend
-
-      - name: Cache Giskard test resources
-        uses: actions/cache@v3
-        with:
-          path: ~/.giskard
-          key: giskard-test-resources
-          restore-keys: giskard-test-resources
-
-      - name: Cache Python dependencies
-        uses: actions/cache@v3
-        with:
-          path: python-client/.venv
-          key: ${{ runner.os }}-client-${{ hashFiles('python-client/pdm.lock')}}
-          restore-keys: ${{ runner.os }}-client
 
       - name: "Python client: extract current package version"
         if: ${{ steps.extract_branch.outputs.branch == 'main' }}
@@ -150,3 +146,5 @@ jobs:
         with:
           name: giskard-dev.whl
           path: python-client/dist/*whl
+
+      - run: echo "$(du -sh $(ls -A ~) | sort -hr)"
