+++ python-client/tests/pytorch/test_sst2_pytorch.py
@@ -1,0 +1,5 @@
+import tests.utils
+
+
+def test_sst2_pytorch_upload(sst2_model, sst2_data):
+    tests.utils.verify_model_upload(sst2_model, sst2_data)
